// src/types/generated/models/PricingPlan.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PlanFeatures } from './PlanFeatures';
export type PricingPlan = {
    billingPeriod: string;
    features: PlanFeatures;
    name: string;
    price: number;
    recommended: boolean;
};